Move Request to internal mongodb package
